requirements.txt
# ==============================================================================
# Core Scientific Computing (For SAM-3 compatibility)
# ==============================================================================
numpy==1.26.4
scipy==1.14.1
pandas==2.2.3
einops

# ==============================================================================
# PyTorch Ecosystem (2.8.0 Series)
# ==============================================================================
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0

# ==============================================================================
# SAM-3 & Meta Research
# ==============================================================================
-e git+https://github.com/facebookresearch/sam3.git#egg=sam3
iopath>=0.1.10
pycocotools>=2.0.8
# onnx>=1.17.0
# onnxruntime-gpu>=1.20.1

# ==============================================================================
# Core Computer Vision & Processing
# ==============================================================================
opencv-python==4.10.0.84
Pillow==10.4.0
decord==0.6.0; sys_platform != "darwin"
timm>=1.0.17
ultralytics

# ==============================================================================
# Others
# ==============================================================================
wandb==0.19.1
python-dotenv==1.0.1
tqdm==4.67.1
matplotlib==3.10.0
gdown
ipykernel